move common code to common lib
