+++ build.sbt
@@ -23,10 +23,15 @@ def setupProject(
     .dependsOn(dependsOn: _*)
     .settings(libraryDependencies ++= externalDependencies)
 }
+lazy val common = setupProject(
+  project,
+  "common",
+  externalDependencies = ServiceDependencies.common)
 
 lazy val ingestor = setupProject(
   project,
   folder = "ingestor",
+  localDependencies = Seq(common),
   externalDependencies = ServiceDependencies.ingestor
 )
