+++ common/src/main/scala/weco/concepts/common/model/Concept.scala
@@ -1,5 +1,5 @@
 // TODO: These will probably end up living in a common project/library in this repo
-package weco.concepts.ingestor.model
+package weco.concepts.common.model
 
 case class Concept(
   identifier: Identifier,
