+++ project/Dependencies.scala
@@ -43,12 +43,12 @@ object ExternalDependencies {
 
 object ServiceDependencies {
   import ExternalDependencies._
+  val common: Seq[ModuleID] = scalatest ++
+    logging ++
+    config ++
+    uPickle
 
   val ingestor: Seq[ModuleID] =
-    scalatest ++
-      logging ++
-      config ++
-      uPickle ++
       Seq(akka.actorTyped, akka.stream, akka.http, akka.streamTestkit)
 
   val aggregator: Seq[ModuleID] = scalatest ++ logging ++ config
